+++ src/core/types/router.ts
@@ -34,7 +34,7 @@ export interface Router {
      * Route the request url string to a controller
      * @return controller relative path to application entry file
      */
-    map(url: URL.Url, method: string): { id: string, filename: string };
+    map(url: URL.Url, method: string): string;
 }
 
 /**