+++ src/core/watch.service.ts
@@ -1,54 +1,66 @@
-"use strict";
-/**
+/***
  * Copyright (c) 2017 [Arthur Xie]
  * <https://github.com/kite-js/kite>
- *
+ * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * 
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
  */
-Object.defineProperty(exports, "__esModule", { value: true });
-const fs = require("fs");
-const path = require("path");
+
+import * as fs from 'fs';
+import * as path from 'path';
+import { LogService } from './log.service';
+
 /**
  * Watch service for watching file (controllers, services, and related files)
  */
-class WatcherService {
+export class WatchService {
+    /**
+     * Watching list
+     */
+    private watching: Map<string, (filename: string) => void>;
+
+    /**
+     * Is watching enabled
+     */
+    private enabled: boolean;
+
+    private node_modules = path.sep + 'node_modules' + path.sep;
+
+    logService: LogService;
+
     /**
      * Watch service for watching file (controllers, services, and related files) changes
-     *
+     * 
      * If service is `enabled`, every file change event will cause a module deletion from NodeJS `require.cache`,
      * the module will be loaded with changed content next time.
-     *
-     * This service will watch for all files recursively, if a file in set to be watched, it's children module
+     * 
+     * This service will watch for all files recursively, if a file in set to be watched, it's children module 
      * will be also watched, except:
      * - files in 'node_modules'
      * - files in Kite's root (in case of someone use Kite as a lib out of NodeJS module)
      */
-    constructor(kiteroot, interval = 2000) {
-        this.kiteroot = kiteroot;
-        this.interval = interval;
-        this.node_modules = path.sep + 'node_modules' + path.sep;
+    constructor(private kiteroot: string, private interval = 2000) {
         // enabled ? this.enable() : this.disable();
     }
+
     /**
      * Enable watching service
      */
-    setEnabled(enabled) {
+    setEnabled(enabled: boolean) {
         if (enabled) {
             this.enabled = true;
             if (!this.watching) {
                 this.watching = new Map();
                 this.logService.info('Watching for file changes');
             }
-        }
-        else {
+        } else {
             this.enabled = false;
             if (this.watching) {
                 this.watching.forEach((callback, filename) => {
@@ -58,12 +70,13 @@ class WatcherService {
             delete this.watching;
         }
     }
+
     /**
      * Watch a file, if change detected, reload it
-     *
+     * 
      * Please note that this watcher is only works with node modules
      */
-    watch(filename, callback) {
+    watch(filename: string, callback?: (filename: string) => void) {
         // skip if:
         // 1. service not enabled
         // 2. file is under "node_modules"
@@ -73,41 +86,51 @@ class WatcherService {
             filename.startsWith(this.kiteroot + path.sep)) {
             return;
         }
+
         // if a file is being watched, update callback function
         if (this.watching.has(filename)) {
             // update callback if set
             if (callback) {
                 this.watching.set(filename, callback);
             }
+
             return;
         }
+
+
         fs.watchFile(filename, { interval: this.interval }, (curr, prev) => {
             this.logService.info(`Change detected at "${filename}"`);
             this.clearCache(filename);
         });
+
         this.watching.set(filename, callback);
+
         if (require.cache[filename]) {
-            require.cache[filename].children.forEach((mod) => {
+            require.cache[filename].children.forEach((mod: NodeModule) => {
                 this.watch(mod.filename);
             });
         }
     }
+
     /**
      * Clear node require cache, it's also search for parent modules, if parent modules are being watched, remove them too
      * @param filename filename to clear
      */
-    clearCache(filename) {
-        let removeList = new Set();
+    private clearCache(filename: string) {
+        let removeList: Set<string> = new Set();
         removeList.add(filename);
-        let keys = Object.keys(require.cache);
-        let findRelated = (filenames) => {
-            let removeWork = new Set();
+
+        let keys: string[] = Object.keys(require.cache);
+
+        let findRelated = (filenames: string[]) => {
+            let removeWork: Set<string> = new Set();
             keys.forEach(key => {
-                let mod = require.cache[key];
+                let mod = require.cache[key] as NodeModule;
                 // Only find managed files
                 if (!this.watching.has(mod.filename) || removeList.has(mod.filename)) {
                     return;
                 }
+
                 mod.children.every(child => {
                     if (filenames.includes(child.filename)) {
                         removeWork.add(mod.filename);
@@ -117,14 +140,18 @@ class WatcherService {
                     return true;
                 });
             });
+
             if (removeWork.size) {
                 findRelated(Array.from(removeWork));
             }
-        };
+        }
+
         findRelated([filename]);
+
         removeList.forEach(fn => {
             let watching = this.watching;
             let callback = this.watching.get(fn);
+
             this.watching.delete(fn);
             delete require.cache[fn];
             fs.unwatchFile(fn);
@@ -133,6 +160,7 @@ class WatcherService {
             }
         });
     }
+
     /**
      * Print out what is being watched
      */
@@ -140,5 +168,3 @@ class WatcherService {
         console.log(this.watching.keys());
     }
 }
-exports.WatcherService = WatcherService;
-//# sourceMappingURL=watcher.service.js.map