+++ src/kite.ts
@@ -23,23 +23,25 @@ import { ErrorService } from './core/error.service';
 import { ControllerFactory } from './core/controller.factory';
 import { ControllerMetadata, getControllerMetadata } from './core/metadata/controller';
 import { getCallerPath } from './core/callsite';
-import { parseSize } from './utils/parse.size';
+import { parseSize } from './utils/parsesize';
 import { HttpRouter } from './utils/http.router';
-import { WatcherService } from './core/watcher.service';
+import { WatchService } from './core/watch.service';
 import { Router } from './core/types/router';
 
 import * as URL from 'url';
 import * as path from 'path';
-import * as querystring from 'querystring';
 import * as cluster from 'cluster';
 import * as os from 'os';
 
 import { DefaultConfig } from './default.config';
-import { Parser, ParserEntity } from './core/types/parser';
+import { Receiver, ReceiverEntity, ReceiverProvider } from './core/types/receiver';
 import { ERROR_CODES } from './core/error.codes';
 import { Middleware } from './core/types/middleware';
 import { Server, createServer, IncomingMessage, ServerResponse } from 'http';
-import { ParserProvider } from './index';
+import { getEntryParams } from './core/metadata/entry';
+import { RequestHandler } from './core/types/request-handler';
+import { ResponseHandler } from './core/types/response-handler';
+import { Provider } from './core/types/provider';
 
 /**
  * Kite 
@@ -47,8 +49,6 @@ import { ParserProvider } from './index';
  * TODO: improve cluster processes
  */
 export class Kite {
-    private static instance: Kite;
-
     private server: Server;
 
     private config: Config;
@@ -63,16 +63,18 @@ export class Kite {
 
     private logService: LogService;
 
-    private watcherService: WatcherService;
+    private watchService: WatchService;
 
-    private parsers: { [name: string]: Parser };
+    private receivers: { [name: string]: Receiver };
 
     private middlewares: Set<Middleware> = new Set();
 
     private router: Router;
 
-    private constructor(workdir: string, config: string | Config = {}) {
-        this.workdir = workdir;
+    private _provider: Provider;
+
+    public constructor(config: string | Config = {}) {
+        this.workdir = getCallerPath();
 
         this.log('Kite framework ver ' + VERSION);
         this.log(`Working at directory ${this.workdir}`);
@@ -93,10 +95,11 @@ export class Kite {
             response.on('error', (err) => {
                 this.logService.error(err);
             });
-            this.requestListener(request, response);
+            this.onRequest(request, response);
         });
 
         server.on('clientError', (err, socket) => {
+            console.error(err);
             socket.end('HTTP/1.1 400 Bad Request\r\n\r\n');
         });
 
@@ -110,21 +113,6 @@ export class Kite {
         });
 
         this.log('Ready to fly');
-    }
-
-    /**
-     * Initialize a Kite instance with given configuration
-     * @param config 
-     */
-    public static init(config: string | Config = {}): Kite {
-        return this.instance || (this.instance = new Kite(getCallerPath(), config));
-    }
-
-    /**
-     * Get current running Kite instance
-     */
-    public static getInstance(): Kite {
-        return this.instance;
     }
 
     /**
@@ -169,13 +157,13 @@ export class Kite {
         this.logService = new LogService(cfg.log.level, cfg.log.out, cfg.log.err);
 
         // set default router
-        if (!cfg.router) {
-            let rootdir = path.join(this.workdir, 'controllers');
-            this.router = new HttpRouter(rootdir);
-        } else if (typeof cfg.router === 'object' && cfg.router.map) {
+        if (typeof cfg.router === 'object' && cfg.router.map) {
             this.router = cfg.router;
         } else if (typeof cfg.router === 'function') {
             this.router = cfg.router.call(null);
+        } else {
+            let rootdir = path.join(this.workdir, 'controllers');
+            this.router = new HttpRouter(rootdir);
         }
 
         // built-in errors
@@ -186,19 +174,19 @@ export class Kite {
         }
 
         // build data hub from factory
-        this.parsers = {};
-        if (cfg.parserProvider) {
-            let parsers: ParserProvider[] = [];
+        this.receivers = {};
+        if (cfg.receiverProvider) {
+            let parsers: ReceiverProvider[] = [];
 
-            if (typeof cfg.parserProvider === 'function') {
-                parsers.push(cfg.parserProvider);
-            } else if (Array.isArray(cfg.parserProvider)) {
-                parsers = cfg.parserProvider;
+            if (typeof cfg.receiverProvider === 'function') {
+                parsers.push(cfg.receiverProvider);
+            } else if (Array.isArray(cfg.receiverProvider)) {
+                parsers = cfg.receiverProvider;
             }
 
             parsers.forEach(provider => {
                 let { contentType, parser } = provider.call(null);
-                this.parsers[contentType] = parser;
+                this.receivers[contentType] = parser;
             });
         }
 
@@ -209,10 +197,10 @@ export class Kite {
 
         this.controllerFactory.logService = this.logService;
 
-        if (!this.watcherService) {
-            this.watcherService = new WatcherService(__dirname);
-            this.watcherService.logService = this.logService;
-            this.controllerFactory.watcherService = this.watcherService;
+        if (!this.watchService) {
+            this.watchService = new WatchService(__dirname);
+            this.watchService.logService = this.logService;
+            this.controllerFactory.watchService = this.watchService;
         }
 
         let oldConfig = this.config;
@@ -220,16 +208,21 @@ export class Kite {
         Object.seal(this.config);
         Object.freeze(this.config);
 
-        // Enable file watch if config.watch is on
-        this.watcherService.setEnabled(Boolean(this.config.watch));
-
         if (typeof config === 'string') {
             this.watchConfigFile(config);
         }
-        // listen again if port / hostname changed
-        if (this.server && (oldConfig.port !== cfg.port || oldConfig.hostname !== cfg.hostname)) {
-            this.fly();
-        }
+    }
+
+    /**
+     * Set watch mode
+     * Watch mode allows you code & test APIs smoothly without restarting Kite server.
+     * 
+     * @param flag - true (default), Kite work in watch mode, suggested in dev mode
+     *             - false, turn off watch suggested set to `false` in production
+     */
+    watch(flag: boolean = true) {
+        this.watchService.setEnabled(flag);
+        return this;
     }
 
     /**
@@ -238,7 +231,7 @@ export class Kite {
      */
     private watchConfigFile(filename: string) {
         // watch for config file changing
-        this.watcherService.watch(filename, (configFilename) => {
+        this.watchService.watch(filename, (configFilename) => {
             this.log('Reload configuration');
             this._init(configFilename);
         });
@@ -247,7 +240,7 @@ export class Kite {
     /**
      * Relase your kite, let it fly
      */
-    fly() {
+    fly(port: number = 4000, hostname: string = 'localhost') {
         if (this.config.workers && cluster.isMaster) {
             this.log(`Master process is running #${process.pid}`);
 
@@ -266,7 +259,7 @@ export class Kite {
             this.server.close();
         }
 
-        this.server.listen(this.config.port, this.config.hostname, () => {
+        this.server.listen(port, hostname, () => {
             let { address, port } = this.server.address();
             this.log(`Flying! server listening at ${address}:${port}`, '\x1b[33m');
         });
@@ -276,63 +269,75 @@ export class Kite {
 
     /**
      * Request listener, process all requests here
-     * @param { http.IncomingMessage } request 
-     * @param { http.ServerResponse } response 
+     * @param { IncomingMessage } request 
+     * @param { ServerResponse } response 
      */
-    private async requestListener(request: IncomingMessage, response: ServerResponse) {
+    private async onRequest(request: IncomingMessage, response: ServerResponse) {
         try {
             let url = URL.parse(request.url, true),
                 inputs = url.query, // URL query string
-                { id, filename } = this.router.map(url, request.method),    // map to actual filename
-                api = await this.controllerFactory.get(id, filename),       // get controller instance
-                // metadata: ControllerMetadata = getControllerMetadata(api.constructor),
-                holder;
+                filename = this.router.map(url, request.method),    // map to actual filename
+                trainData;
+                // api = await this.controllerFactory.get(filename);       // get controller instance
+            // metadata: ControllerMetadata = getControllerMetadata(api.constructor),
 
-            // if there is any message-body sent from client, try to parse it
-            // an entity-body is explicitly forbidden in TRACE, and ingored in GET
-            if ((request.headers['content-length'] || request.headers['transfer-encoding']) &&
+            let controller = this.controllerFactory.getController(filename);
+            if (this._provider) {
+                trainData = this._provider.exec(request, controller, inputs);
+            }
+            let api: any = await this.controllerFactory.getInstance(controller, trainData);
+
+            // if api handle http request it self, skip "input" parsing
+            if ((api as RequestHandler).onRequest) {
+                inputs = await (api as RequestHandler).onRequest(request, url.query);
+            } else if ((request.headers['content-length'] || request.headers['transfer-encoding']) &&
                 request.method !== 'GET' &&
                 request.method !== 'TRACE') {
+                // if there is any message-body sent from client, try to parse it
+                // an entity-body is explicitly forbidden in TRACE, and ingored in GET
 
-                let contentType = <string>request.headers['content-type'] || '',
+                let contentType = <string>request.headers['content-type'] || 'text/plain',
                     entityBody = await this.getEntityBody(request);
 
-                if (this.parsers[contentType]) {
+                if (this.receivers[contentType]) {
                     try {
-                        let data = this.parsers[contentType](entityBody);
+                        let data = this.receivers[contentType](entityBody);
                         inputs = Object.assign({}, url.query, data);
                     } catch (e) {
                         this.logService.error(e);
                         throw new KiteError(1010);
                     }
                 } else {
-                    this.logService.warn(`Unsupport content type "${contentType}"`);
-                    inputs = entityBody;
+                    this.logService.warn(`Unsupported content type "${contentType}"`);
+                    // inputs = entityBody;
+                    inputs.$data = entityBody;
                 }
             }
 
             // Call middlewares
             let middleResult: any;
+
             for (let middleware of this.middlewares) {
-                middleResult = await middleware.call(null, response, request, api, inputs);
-                // is it a HolderClass ?
-                if (this.config.holderClass && middleResult instanceof this.config.holderClass) {
-                    holder = middleResult;
-                } else if (middleResult === false) {
+                middleResult = middleware.exec(request, response, api, inputs);
+                // if return type is promise then wait promise return
+                if (middleResult instanceof Promise) {
+                    middleResult = await middleResult;
+                }
+
+                // ends the response if middleware explicitly returns false, else continue
+                if (middleResult === false) {
                     response.end();
                     return;
                 }
             }
 
-            // call API with pre-generated $proxy(inputs, holder, context)
-            let result = await api.$proxy(
-                inputs,
-                holder,
-                request,
-                response
-            );
-            // if api havn't write response, call responder to output contents
-            if (!response.headersSent) {
+            // call API with pre-generated $proxy(inputs)
+            let result = await api.$proxy(inputs, request);
+
+            // let this controller handle all response if "handleResponse" function is available
+            if ((api as ResponseHandler).onResponse) {
+                (api as ResponseHandler).onResponse(response, result);
+            } else {
                 this.config.responder.write(result, response);
             }
         } catch (err) {
@@ -347,12 +352,10 @@ export class Kite {
                 } catch (err) {
                     this.logService.error(err);
                     let error = this.errorService.getError(1001);
-                    response.write(JSON.stringify({ error }));
+                    response.end(JSON.stringify({ error }));
                 }
             }
         }
-
-        response.end();
     }
 
     /**
@@ -400,9 +403,14 @@ export class Kite {
     }
 
     /**
-     * Get config of current Kite instance
+     * Set train kite data provider.
+     * 
+     * If a provider is given, Kite will invoke the `exec()` method when request comes, 
+     * the return value must be a "injectable" object
+     * @param provider Provider instance
      */
-    getConfig() {
-        return this.config;
+    provider(provider: Provider) {
+        this._provider = provider;
+        return this;
     }
 }