+++ src/utils/index.ts
@@ -14,5 +14,5 @@
  */
 
 export * from './http.router';
-export * from './xform.parser.provider';
-export * from './json.parser.provider';
+export * from './xform-receiver.provider';
+export * from './json-receiver.provider';